Figured categories and menu out
So I figured the categories and menu out. Categories can be set with
some code to get posts of that type, got excerpts to appear where I want
them. New category pages have been copied out but not actually edited
and properly set yet, these are technical and personal.

An All posts page was also created and split out from the home page of
the blog.

To change the items in the menu at the top all that needs to be done is
to edit the files in _config.yml